Changed himmel, rosen, dejong to use doubles for their internal OF calculations even though they take floats as input. Hoping to avoid some rounding.

+++ SGA/src/HimmelblauFunction.java
@@ -10,8 +10,8 @@ public class HimmelblauFunction implements Problem {
 	public double fitness(boolean[] candidate) {
 		boolean[] xBool = Arrays.copyOf(candidate, 32);
 		boolean[] yBool = Arrays.copyOfRange(candidate, 32, 64);
-		float x = OFHelper.floatFromBoolArray(xBool);
-		float y = OFHelper.floatFromBoolArray(yBool);
+		double x = OFHelper.floatFromBoolArray(xBool);
+		double y = OFHelper.floatFromBoolArray(yBool);
 		return Math.pow(x * x + y - 11, 2) + Math.pow(x + y * y - 7, 2);
 	}
